+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The polish dictionary".
 - Tests for "1010(1)".
 - Tests for "Vote counting".
+- Tests for "Maze for the champions".
 
 ## [1.16.0] - 2022-12-31
 ### Added